refactor(python): Move inference code out of executor module
